Fix CORS, update config & requirements for segmentation service

+++ backend/app/api/schemas/segmentation.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 from typing import Optional
 
 class PredictionResponse(BaseModel):
-    panel_area: int
+    panel_area: float
     panel_count: int
     total_power_kwp: float
     annual_energy_mwh: float
